feat: Add Star Wars character retrieval script
This commit adds a script that retrieves and prints the characters of a Star Wars movie based on the movie ID provided as a command-line argument. The script uses the Star Wars API and the `request` module to make HTTP requests and display the character names in the same order as the "characters" list in the API response.